test_cli: Change listener back to forked process
This reverts the listener portion of
bea5c925b860ab4bc4ce0c72e161a31592e1b1dc ("CLI test improvements")
but changes it to use subprocess instead of os.fork and os.execv.
Turns out any use of Python code within unittest after os.fork is
dangerous, and the previous attempt to run the listener in a fork
just happened to work on Linux pollers, but would fail in odd
ways on OS X / FreeBSD.  To be clear, this just affected the test
harness; the actual executables were fine.

I'm not happy with going back to launching a listener as a
subprocess, waiting 3 seconds and hoping it's started by then,
but it looks like running a listener directly within unittest
cannot be made safe.