Fix Windows test runner issues, hopefully for the last time
So, the intial problem was:
  [WinError 10013] An attempt was made to access a socket in a way
  forbidden by its access permissions

Some initial searching pointed to Azure having an undocumented
outbound bind limit.  Makes sense to discourage outbound port
scanning, but seemed weird since I was binding to 127.0.0.1 and
connecting to 127.0.0.1.  Nonetheless, I figured it was flood and
adaptive, so I disabled them on Windows Github runners, and it
seemed to initially go away, but then them back.

I then figured it was something to do with listening on certain
high ports, so put in a check to verify we could bind to a port
before passing it on to the test, rather than blindly picking
between 49152 and 65535.  Again, that seemed to work, so I pushed,
but it came back originally.

Turns out I had forgotten the actual bind and yes, that original
error is indicative of many possible problems, including existing
bound port.  On the Github Windows runners, some process under
iphlpsvc is binding to a series of 200 UDP ports in sequence,
beginning at a random high port.  It was just chance that the
above two attempts seemed to succeed; none of our tests happened
to pick one of those 200 ports.

So to summarize: Nothing special about the Windows runners, except
that there's something binding to 200 UDP ports.  The code in
place should (hopefully) correctly pick a random port to bind to
which is not already in use.

If it fails after this, I may need to take a walk around the
block.

Changes:
- Raise if --debug and error during setup_sockets
- Remove test_adaptive and test_flood skips
- Actually bind during port check
- Warn UserWarning if it took multiple attempts to get a port